fixed product edit in development environment
