+++ packages/framework/assets/js/admin/validation/form/index.js
@@ -11,3 +11,4 @@ import './validationPromoCode';
 import './validationHreflangSetting';
 import './validationStore';
 import './validationTransportPrices';
+import './validationUniqueProductCatnum';
