+++ packages/framework/src/Resources/translations/validators.cs.po
@@ -457,6 +457,9 @@ msgstr "Název produktu nesmí být delší než {{ limit }} znaků"
 msgid "Product parameters are duplicate."
 msgstr "Parametry produktů jsou duplicitní."
 
+msgid "Product with entered catalog number already exists"
+msgstr "Produkt s tímto katalogovým číslem již existuje"
+
 msgid "Promo code \"%invalidPromoCode%\" is not available for pricing group ID \"%pricingGroupId%\"."
 msgstr "Slevový kupón \"%invalidPromoCode%\" není dostupný pro Vaši cenovou skupinu."
