+++ project-base/app/assets/js/admin/validation/index.js
@@ -1,2 +1,1 @@
 import './promoCodeValidator';
-import './uniqueProductCatnumValidator';
