+++ project-base/app/translations/validators.en.po
@@ -7,9 +7,6 @@ msgstr ""
 msgid "An array of values is not supported as an input for \"slider\" type parameter. Use \"minimalValue\" and \"maximalValue\" instead."
 msgstr ""
 
-msgid "Catalog number cannot be longer than {{ limit }} characters"
-msgstr ""
-
 msgid "Email cannot be longer than {{ limit }} characters"
 msgstr ""
 
@@ -77,9 +74,6 @@ msgid "Product prefix name cannot be longer than {{ limit }} characters"
 msgstr ""
 
 msgid "Product suffix name cannot be longer than {{ limit }} characters"
-msgstr ""
-
-msgid "Product with entered catalog number already exists"
 msgstr ""
 
 msgid "Provided hash is not valid."
