+++ upgrade-notes/backend_20241030_090019.md
@@ -1,0 +1,6 @@
+#### fix product edit in dev environment ([#3544](https://github.com/shopsys/shopsys/pull/3544))
+
+-   [features moved](#movement-of-features-from-project-base-to-packages) from project-base to the framework package:
+    -   product catnum is now required and forced to be unique
+-   check your code for use of the `FormBuilderInterface::getAttributes()` method and replace it with `FormBuilderInterface::getOptions()['attr'] ?? []` to prevent errors about array-to-string conversion
+-   see #project-base-diff to update your project
